Fix moving the splitter
Without count it should be exactly one.

+++ src/status.h
@@ -91,7 +91,7 @@ typedef struct
 
 	int scroll_bind_off;
 	enum Split split;
-	float splitter_pos;
+	int splitter_pos;
 }Status;
 
 extern Status curr_stats;
